--- GoodsStore/GoodsStore.Client.ViewModels/GoodsStore.Client.ViewModels.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Net.Http.Json" Version="5.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\GoodsStore.Business.Models\GoodsStore.Business.Models.csproj" />
-  </ItemGroup>
-
-</Project>
