+++ GoodsStore/GoodsStore.Client/VIewModels/Abstract/IGenericCollectionVM.cs
