+++ GoodsStore/GoodsStore.Client/VIewModels/Abstract/IGenericItemVM.cs
