+++ GoodsStore/GoodsStore.Client/VIewModels/Concrete/GenericCollectionVM.cs
