+++ GoodsStore/GoodsStore.Client/VIewModels/Concrete/GenericItem.cs
